requirements.txt
torch>=1.8.1
Pillow>=8.1.0
numpy>=1.20.1
seldon-core==1.17.1
torchvision